pyproject.toml
[project]
name = "strands-agents-bugs"
version = "0.1.0"
authors = [{ name = "Qian Zhang", email = "contact@qzsh.email" }]
requires-python = ">=3.13, <3.14"
dependencies = [
  "strands-agents[gemini]>=1.16.0",
  "strands-agents-tools>=0.2",
  "python-dotenv",
]

[dependency-groups]
dev = ["pre-commit>=4.0", "ruff>=0.11"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strands_agents_bugs"]

[project.scripts]
gemini_missing_tool_id = "strands_agents_bugs.gemini_missing_tool_id:main"
